chore: moving router to a single file
