--- src/servers/authorization/router.ts
@@ -1,2 +1,0 @@
-import express from "express"
-export const authRouter = express.Router({ caseSensitive: true })
